requirements.txt
arrow==0.15.7
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
idna==2.9
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
